x/backend/requirements.txt
Flask
flask-cors
tweepy
python-dotenv
sentence-transformers
torch
numpy
dateparser
pandas
